+++ app/Http/Controllers/TicketController.php
@@ -12,6 +12,7 @@ class TicketController extends Controller
 {
     public function create(TicketRequest $request, Profile $profile): JsonResponse
     {
+        $this->authorize('createOrModify', [Ticket::class, $profile]);
         $profile->ticket()->updateOrCreate([
             'camera' => $request->camera,
             'description' => $request->description
@@ -28,11 +29,13 @@ class TicketController extends Controller
 
     public function show(Profile $profile): JsonResource
     {
+        $this->authorize('view', [Ticket::class, $profile]);
         return new TicketResource($profile->ticket());
     }
 
     public function update(TicketRequest $request, Profile $profile): JsonResponse
     {
+        $this->authorize('createOrModify', [Ticket::class, $profile]);
         $profile->ticket()->update([
             'camera' => $request->camera,
             'description' => $request->description
@@ -49,6 +52,7 @@ class TicketController extends Controller
 
     public function destroy(Profile $profile): void
     {
+        $this->authorize('createOrModify', [Ticket::class, $profile]);
         $profile->ticket()->delete();
     }
 }
